# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Roster Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all user accounts
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /squads:
    get:
      operationId: listSquads
      tags: [squads]
      summary: List all squads
      responses:
        "200":
          description: List of squads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Squad"
    post:
      operationId: createSquad
      tags: [squads]
      summary: Create a squad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSquadRequest"
      responses:
        "201":
          description: Created squad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Squad"
  /squads/{id}:
    put:
      operationId: updateSquad
      tags: [squads]
      summary: Update a squad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSquadRequest"
      responses:
        "200":
          description: Updated squad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Squad"
    delete:
      operationId: deleteSquad
      tags: [squads]
      summary: Delete a squad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /roster:
    get:
      operationId: listMembers
      tags: [roster]
      summary: List all roster members
      responses:
        "200":
          description: List of roster members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RosterMember"
    post:
      operationId: createMember
      tags: [roster]
      summary: Add a member to the roster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberRequest"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterMember"
  /roster/{id}:
    get:
      operationId: getMember
      tags: [roster]
      summary: Get a roster member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Roster member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterMember"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMember
      tags: [roster]
      summary: Update a roster member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterMember"
    delete:
      operationId: deleteMember
      tags: [roster]
      summary: Remove a member from the roster
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, viewer]
        mfaEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
      required: [id, username, email, role, mfaEnabled, createdAt]
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, viewer]
      required: [username, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, viewer]
        password:
          type: string
...
    Squad:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        memberCount:
          type: integer
      required: [id, name]
    CreateSquadRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
    UpdateSquadRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    RosterMember:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
          nullable: true
        rankId:
          type: integer
          nullable: true
        rankName:
          type: string
          nullable: true
        rankAbbreviation:
          type: string
          nullable: true
        squadId:
          type: integer
          nullable: true
        squadName:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, suspended]
        notes:
          type: string
          nullable: true
        joinedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, username, status, joinedAt, updatedAt]
    CreateMemberRequest:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        rankId:
          type: integer
        squadId:
          type: integer
        status:
          type: string
          enum: [active, inactive, suspended]
        notes:
          type: string
      required: [username]
    UpdateMemberRequest:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        rankId:
          type: integer
          nullable: true
        squadId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, inactive, suspended]
        notes:
          type: string
...