# lib/api-zod/src/generated/types/createMemberRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Roster Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMemberRequestStatus } from "./createMemberRequestStatus";

export interface CreateMemberRequest {
  username: string;
  displayName?: string;
  rankId?: number;
  squadId?: number;
  status?: CreateMemberRequestStatus;
  notes?: string;
}