# lib/api-zod/src/generated/types/rosterMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Roster Management API
 * OpenAPI spec version: 0.1.0
 */
import type { RosterMemberStatus } from "./rosterMemberStatus";

export interface RosterMember {
  id: number;
  username: string;
  displayName?: string | null;
  rankId?: number | null;
  rankName?: string | null;
  rankAbbreviation?: string | null;
  squadId?: number | null;
  squadName?: string | null;
  status: RosterMemberStatus;
  notes?: string | null;
  joinedAt: Date;
  updatedAt: Date;
}