# lib/api-zod/src/generated/types/squad.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Roster Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Squad {
  id: number;
  name: string;
  description?: string | null;
  memberCount?: number;
}